extensions/serial-input-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "serial-input-extension"
version = "0.1.0"
description = "Neutralino serial input extension"
authors = [ { name = "Variable Watcher" } ]
requires-python = ">=3.9"
dependencies = [
  "pyserial>=3.5",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
